move helper functions into dibs.utils.func
